Move pick_channels_for_modalities to _channel_pick module, add whats_new
Extract helper to moabb/datasets/_channel_pick.py to break circular
import chain (base → bids_interface/preprocessing → utils → base).
All imports are now global. Re-export from utils for public API.
Add whats_new entries for the return_all_modalities feature and bug fixes.
