+++ moabb/datasets/preprocessing.py
@@ -8,6 +8,8 @@ import numpy as np
 from sklearn.base import BaseEstimator, TransformerMixin
 from sklearn.pipeline import FunctionTransformer, Pipeline, _name_estimators
 
+from moabb.datasets._channel_pick import pick_channels_for_modalities
+
 
 # Handle different scikit-learn versions for _VisualBlock import
 # sklearn >= 1.6 moved _VisualBlock to sklearn.utils._repr_html.estimator
@@ -757,8 +759,6 @@ class RawToEpochs(FixedTransformer):
             raise ValueError("raw must be a mne.io.BaseRaw")
 
         if self.channels is None:
-            from moabb.datasets.utils import pick_channels_for_modalities
-
             picks = pick_channels_for_modalities(raw.info, self.return_all_modalities)
         else:
             available_channels = raw.info["ch_names"]
