+++ moabb/datasets/utils.py
@@ -24,6 +24,7 @@ from moabb.analysis.plotting import (
     dataset_bubble_plot,
     get_dataset_area,
 )
+from moabb.datasets._channel_pick import pick_channels_for_modalities  # noqa: F401
 from moabb.datasets.base import BaseDataset
 from moabb.utils import aliases_list
 
@@ -32,14 +33,6 @@ logger = logging.getLogger(__name__)
 
 dataset_list = []
 dataset_dict = {}
-
-
-def pick_channels_for_modalities(info, return_all_modalities=False):
-    """Pick channel indices: all non-stim if *return_all_modalities*, else EEG only."""
-    if return_all_modalities:
-        stim_picks = set(mne.pick_types(info=info, stim=True))
-        return [i for i in range(len(info.ch_names)) if i not in stim_picks]
-    return mne.pick_types(info=info, eeg=True, stim=False)
 
 
 def _init_dataset():
